% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggTraj.R
\name{ggTraj}
\alias{ggTraj}
\title{Plot animal trajectories (ggplot)}
\usage{
ggTraj(df, xyi = c("x", "y", "ndowid"))
}
\arguments{
\item{df}{The dataframe of the animals GPS points}

\item{xyi}{A character vector of the x, y coordinates and animal ID. Default
values are c('x', 'y', 'ndowid')}
}
\value{
A ggplot of the animal trajectory. The plot has basic styling applied for
consistency. The plot can be saved to a variable and styling can be added with ggplot2 functions.
}
\description{
Plot animal trajectories (ggplot)
}
\examples{
ggTraj(df, c('x', 'y', 'ndowid'))

}
\author{
Mitchell Gritts
}

