% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveParams.R
\name{moveParams}
\alias{moveParams}
\title{Estimate movement parameters from timestamped xy data}
\usage{
moveParams(x, y, timestamp, dat = NULL, isPOSIXct = TRUE)
}
\arguments{
\item{x}{the x component of the xy data}

\item{y}{the y component of the xy data}

\item{timestamp}{the timestamp of the xy data}

\item{dat}{additional data for the returned data to be added to (optional)}

\item{isPOSIXct}{whether or not the timestamp data is of class \code{POSIXct}
the default value is \code{TRUE}}
}
\value{
a \code{data.table} with the following columns:
\item{dist}{the distance between xy locations (meters)}
\item{nsd}{net squared displacement (meters)}
\item{dt}{duration between xy locations (seconds)}
\item{speed}{velocity between xy locations (km/hr)}
\item{phi, theta}{absolute and turning angle (respectively)}
\item{vp}{persistence velocity}
\item{vt}{turning velocity}
}
\description{
The function will estimate common movement parameters from the given xy coordinates and
timestamp data. The data can be returnded by itself as a data.table or added to an
existing data.table.
}

