% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/muldDat.R
\docType{data}
\name{muldDat}
\alias{muldDat}
\title{A short MULD trajectory}
\format{A data.table with 500 rows and 3 variables:
\describe{
   \item{timestamp}{chr string of the timestamp}
   \item{long_x}{longitude of the location (WGS 84)}
   \item{lat_y}{latitude of the location (WGS 84)}
   }}
\usage{
muldDat
}
\description{
A dataset containing a random sample of a MULD trajectory of 500 locations.
}
\keyword{datasets}

