% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot3DTraj.R
\name{plot3DTraj}
\alias{plot3DTraj}
\title{3D plot of animal trajectory}
\usage{
plot3DTraj(x, y, z, colval)
}
\arguments{
\item{x}{The x geographic coordinate to plot}

\item{y}{The y geographic coordinate to plot}

\item{z}{The z value for the z axis}

\item{colval}{The value you want to plot as points along the trajectory}
}
\value{
A 3D plot of the the animal trajectory. Each point is connected by a
line. The color of the points is value of the colval argument. Purple is for low
values, Yellow is for high levels.
}
\description{
3D plot of animal trajectory
}
\note{
A good use for the 3D plot is to construct a spacetime cube. In a spacetime cube
the z-axis is the timestamp of the points. There is potential to visualize seasonal use
and revisitation rates. It is useful to set the \code{colval} argument equal to the
timestamp as well to generate a color ramp for the timestamp. The Purple will be for the first
XY locations, Yellow for the last.
}
\examples{
plot3DTraj(df$X, df$Y, df$timestamp, df$timestamp)

}
\author{
Mitchell Gritts
}

