% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTraj.R
\name{plotTraj}
\alias{plotTraj}
\title{Plot an animal trajectory}
\usage{
plotTraj(x, y)
}
\arguments{
\item{x}{The x geographic coordinate to plot}

\item{y}{the y geographic coordinate to plot}
}
\value{
A plot of the the animal trajectory. Each point is connected by a
line. The start of the trajectory is shown as the green circle, the end
as the red circle.
}
\description{
Plot an animal trajectory
}
\examples{
plot_traj(df$X, df$Y)
}

