% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xyConv.R
\name{xyConv}
\alias{xyConv}
\title{Convert xy coordinates from one projection to another.}
\usage{
xyConv(df, xy = c("long_x", "lat_y"), CRSin = "+proj=longlat",
  CRSout = "+proj=utm +zone=11", outclass = "data.table")
}
\arguments{
\item{df}{A dataframe that contains x, y coordinates.}

\item{xy}{A vector of the names for the input x and y coordinates. Default is long_x and lat_y}

\item{CRSin}{The CRS string of the input coordinates, defaults to longlat}

\item{CRSout}{The CRS string of the output coordinates, defaults to UTM Zone 11}

\item{outclass}{The class of the data object to be returned. Options include \code{'data.frame'},
\code{'data.table'}, and \code{'spdf'}. \code{'spdf'} returns an object of class
\code{sp::SpatialPointsDataFrame}.}
}
\value{
the original \code{data.table}, \code{data.frame} or \code{SpatialPointsDataFrame}
 with converted XY coordinates added to the object.
}
\description{
Takes a dataframe with XY coordinates as input and converts the coordinates to
another projection system. The function defaults to converting longlat to UTM zone 11,
however any properly formated CRS string is acceptable for input.
}
\examples{
df <- xyConv(dat, xy = c('geox', 'geoy'))
}
\author{
Mitchell Gritts
}

