% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveParams.R
\name{moveParams}
\alias{moveParams}
\title{Estimate movement parameters from timestamped xy data}
\usage{
moveParams(df, xyti = c("x", "y", "timestamp", "ndowid"), format = NULL)
}
\arguments{
\item{format}{Format of the date time string to be passed to the \code{as.POSIXct}.
The default is \code{NULL} assuming that timestamps will be in the "standard unambiguous format",
i.e. YYYY-MM-DD HH:MM:SS. If the timestamp is already of class \code{POSIXct}, this isn't required.}

\item{x}{The x component of the xy data}

\item{y}{The y component of the xy data}

\item{timestamp}{The timestamp of the xy data}

\item{dat}{Additional data for the returned data to be added to (optional)}
}
\value{
a \code{data.table} with the following columns:
\item{dist}{the distance between xy locations (meters)}
\item{nsd}{net squared displacement (meters)}
\item{dt}{duration between xy locations (seconds)}
\item{speed}{velocity between xy locations (km/hr)}
\item{phi, theta}{absolute and turning angle (respectively)}
\item{vp}{persistence velocity}
\item{vt}{turning velocity}
}
\description{
The function will estimate common movement parameters from the given xy coordinates and
timestamp data. The data can be returnded by itself as a data.table or added to an
existing data.table.
}
\examples{
df <- moveParams(df$X, df$Y, df$timestamp, dat = df, format = '\%Y-\%m-\%d \%H:\%M:\%S')
}
\author{
Mitchell Gritts
}

